refactor: inline Table into mutable_buffer::chunk::Chunk
